Update IdleFarmer.js
I updated the way adding rows and columns works, and they are now dependent on the array itself rather than variables that keep track of the array length, making for cleaner code. Also the main this is that I updated plant, and now the "plant" event propagates to the game object and adds the crops to a list in the game object for future reference and statistics. This will serve as the basis for the hiring mechanic.